+++ test/index_test.rb
@@ -85,7 +85,9 @@ class IndexTest < Minitest::Test
   def test_record_not_found
     store_names ["Product A", "Product B"]
     Product.where(name: "Product A").delete_all
-    assert_search "product", ["Product B"]
+    assert_output nil, /\[searchkick\] WARNING: Records in search index do not exist in database/ do
+      assert_search "product", ["Product B"]
+    end
   ensure
     Product.reindex
   end
@@ -93,7 +95,8 @@ class IndexTest < Minitest::Test
   def test_bad_mapping
     Product.searchkick_index.delete
     store_names ["Product A"]
-    assert_raises(Searchkick::InvalidQueryError) { Product.search "test" }
+    error = assert_raises(Searchkick::InvalidQueryError) { Product.search "test" }
+    assert_equal "Bad mapping - run Product.reindex", error.message
   ensure
     Product.reindex
   end